+++ crates/polars-expr/src/dispatch/mod.rs
@@ -303,7 +303,7 @@ pub fn function_expr_to_udf(func: IRFunctionExpr) -> SpecialEq<Arc<dyn ColumnsUd
         F::Repeat => map_as_slice!(misc::repeat),
         #[cfg(feature = "rank")]
         F::Rank { options, seed } => map!(misc::rank, options, seed),
-        F::AsList => map_as_slice!(list::as_list),
+        F::AsList => map_as_slice!(misc::as_list),
         #[cfg(feature = "dtype-struct")]
         F::AsStruct => {
             map_as_slice!(misc::as_struct)
